Create make_plate.py
Given a list of sample id in one column raw text file, arrange the samples into 8x12 plates. Each cell fills in a sample. Raw is indicated by a, b, c, ..., h. Column is indicated by number 1 to 12. When filling the samples, raw is first. Plate is indexed like "Plate1", "Plate2"....

Two output files are generated. One is the map of the plate showing each cell in each plate. The other is showing the position of each sample by row.